requirements.txt
django==3.2.5
django-import-export==2.5.0

